Define each authenticator individually
Defining them together, even though they technically use the same route in
Conjur, was proving tricky because they each have different requirements for
request bodies, etc. In this commit we split the service authenticators into
separate paths in the spec.

The commit also updates the `rotate_api_key` route to hardcode the authenticator
as `authn` in the route - though the server currently allows requests to come
with other `authn` values, the behavior is the same regardless of what is sent.