Update gh action helm install strategy
Use the official helm/kind-action to create the KinD cluster instead of
installing KinD with curl.

This commit also revises the action strategy for overwriting C# tests.

C# clients include autogenerated tests, but the logic we were using to replace
these with manually written tests was failing in actions in some cases.

The logic is updated in this commit to copy the manual tests into the generated
directory first, and then to stand up the test Docker container.