Updated spec and integration tests for new client
The upgrade to openapi-generator version 5.1.0 caused
changes in how the generated client works that required updates
to the integration tests and spec itself. Namely that making
a request with http info (returning the status code and headers
along with response body) is now carried out by passing a variable
rather than calling a version of the method named `<method>_with_http_info`.
Unfortunately newer versions of the generator do not fully support
the minLength and maxLength fields in OpenAPI so they had to be removed.

The new Python generated client has much better type
checking so some bad types that had previously gone
undetected have been updated. The new client also has
better enum checking, so in certain cases it is no
longer possible to pass a bad value to Conjur in
order to evoke a 404 error. These tests have either
been removed if redundant or replaced with a test ensuring
that the client errors on a badly passed parameter.