Fix and reduce coverage targets
Coverage targets are thresholds with the format "healthy, unhealthy, unstable"
so each threshold moving from left to right should be equal to or less than the
previous one.

I'm also reducing the line coverage threshold because it's currently failing on
a commit that only made changes to the CHANGELOG, so cobertura may be analyzing
more files than it should be.